CLEANUP: reorganize export of cryptohash
